requirements.txt
fastapi
uvicorn
requests
python-dotenv
snowflake-snowpark-python
snowflake-ml-python